Version 26.08.1, Rust SDK 26.08.13, Kotlin 2.3.11, metro 1.4.1, roborazzi 1.71.0. Most of
what came across is a documentation sweep adding KDoc to the API interfaces, renovate bumps
and translations; a hundred or so commits carry actual behaviour.

Six things needed deciding.

The map tile server moved into the Element wellknown content, so the builder comes from
rememberStaticMapBuilder(mapTilerConfig) rather than being constructed. Ours keeps the lazy
load on top: a static map is still only fetched once the person asks for it, or when media
auto-load says otherwise.

They removed the legacy web-based account creation for matrix.org, taking CreateAccountNode
with it. The only thing of ours in that file was routing its external link through the
mxtr-aware custom tab, and that helper is used in nine other places, so the file goes.

Their settings screen gained a Box wrapper for the animated developer entry point, which
landed on top of our multi-account section. Both stay.

The video bitrate cap arrived from their side too, in Int rather than Long. Ours had the
same idea and something more: a source already inside the preset is remuxed rather than
re-encoded. Their types, our remux.

scheduleDatabaseVacuum() moved up into the session setup on their side, so the copy at the
end of setupUserProfile would have run it twice. Dropped. The search backfill next to it is
ours and stays.

RustMatrixClientFactory takes a ContentScannerUrlProvider.Factory now instead of a provider.
Their new RustTemporaryMatrixClient tests do not know about our certificate pinning, so they
needed userCertificatesProvider passing in.

The 149 conflicting screenshots were upstream files lost here during the old rebase repair.
Taken from upstream.

+++ libraries/mediaupload/impl/src/main/kotlin/io/element/android/libraries/mediaupload/impl/VideoCompressorConfig.kt
@@ -34,7 +34,7 @@ internal object VideoCompressorConfigFactory {
         val newFrameRate = min(originalFrameRate, DEFAULT_FRAME_RATE)
 
         // If we need to resize the video, we also want to recalculate the bitrate
-        val optimalBitrate = resizer.calculateOptimalBitrate(Size(width, height), newFrameRate)
+        val optimalBitrate = resizer.calculateOptimalBitrate(Size(width, height), newFrameRate).toInt()
 
         // METADATA_KEY_BITRATE covers the whole container, video plus audio, so it can only
         // over-estimate the video track. That still makes it a safe ceiling: encoding above what
@@ -42,7 +42,7 @@ internal object VideoCompressorConfigFactory {
         // clip filmed at 720p/1.5Mbps comes out of the STANDARD preset at 2.8Mbps, so the
         // "compression" step nearly doubles the upload.
         val sourceBitrate = metadata?.bitrate?.takeIf { it > 0 }
-        val newBitrate = sourceBitrate?.let { min(optimalBitrate, it) } ?: optimalBitrate
+        val newBitrate = sourceBitrate?.let { min(optimalBitrate.toLong(), it).toInt() } ?: optimalBitrate
 
         // Nothing to scale down and nothing to shave off the bitrate: a re-encode would only cost
         // time and quality. Ask for a remux instead, which still runs through the muxer that strips
@@ -54,7 +54,7 @@ internal object VideoCompressorConfigFactory {
 
         return VideoCompressorConfig(
             videoCompressorHelper = resizer,
-            newBitRate = newBitrate.toInt(),
+            newBitRate = newBitrate,
             newFrameRate = newFrameRate,
             canRemux = canRemux,
         )
